added admin exception
Admin is automatically an approved user

+++ src/com/example/backend/User.java
@@ -91,6 +91,10 @@ public class User {
         setAdmin(admin);
         // setAdmin(!DATABASE.hasAdminUser() || admin);
         setLastLogin(LocalDate.now());
+        if(admin){
+            DATABASE.addUser(this);
+            return true;
+        }
         if(PENDING_DATABASE.userExists(name)){
         DATABASE.addUser(this);
         PENDING_DATABASE.deleteUser(username, password);
